services/core-api/requirements.txt
# CRITICAL: motor + pymongo + beanie must stay at these versions — upgrading any one of them without checking compatibility will break the app.

# FastAPI framework for building async APIs
fastapi

# ASGI server for running FastAPI with async/await support and optional performance features (uvloop, httptools)
uvicorn[standard]

# Motor 3.3.2 - async MongoDB driver (PINNED - do not upgrade without testing beanie and pymongo compatibility)
motor==3.3.2

# PyMongo 4.6.3 - synchronous MongoDB driver (PINNED - must match motor version compatibility)
pymongo==4.6.3

# Beanie 1.25.0 - async ODM for MongoDB (PINNED - must be compatible with motor==3.3.2)
beanie==1.25.0

# JWT authentication with cryptography support
python-jose[cryptography]

# Password hashing library with bcrypt support
passlib[bcrypt]

# Multipart form data parsing for file uploads and form submissions
python-multipart

# Groq API client for AI/LLM integration
groq

# ChromaDB - vector database for embeddings and semantic search
chromadb

# Redis client for caching and session management
redis

# Environment variable loader for configuration management
python-dotenv

# Data validation with email validation support
pydantic[email]

# Pydantic settings management for configuration
pydantic-settings

# Async HTTP client for making external API requests
httpx

# Google Calendar OAuth and API client
google-auth
google-auth-oauthlib
google-api-python-client

# Google Gemini API client for real-time transcription
google-genai

# Numpy for audio PCM conversion
numpy